# backend/src/main/resources/openapi/pedidos-api.yaml
openapi: 3.0.3
info:
  title: TPVGallega - API de pedidos
  description: Contrato de la API REST para la gestion de comandas de TPVGallega.
  version: 1.0.0

servers:
  - url: http://localhost:8080

tags:
  - name: Pedidos
    description: Alta, consulta y cambio de estado de comandas

paths:
  /api/pedidos:
    post:
      operationId: crearPedido
      tags: [Pedidos]
      summary: Crea un nuevo pedido (queda en estado PENDIENTE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearPedidoRequest"
      responses:
        "201":
          description: Pedido creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "400":
          description: Solicitud invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listarPedidos
      tags: [Pedidos]
      summary: Lista pedidos, opcionalmente filtrados por estado
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EstadoPedido"
      responses:
        "200":
          description: Listado de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PedidoResponse"

  /api/pedidos/{id}:
    get:
      operationId: obtenerPedido
      tags: [Pedidos]
      summary: Obtiene un pedido por id
      parameters:
        - $ref: "#/components/parameters/PedidoId"
      responses:
        "200":
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "404":
          description: No existe un pedido con ese id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/pedidos/{id}/estado:
    patch:
      operationId: cambiarEstado
      tags: [Pedidos]
      summary: Cambia el estado de un pedido
      parameters:
        - $ref: "#/components/parameters/PedidoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CambiarEstadoRequest"
      responses:
        "200":
          description: Pedido actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "400":
          description: Transicion de estado invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No existe un pedido con ese id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    PedidoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    EstadoPedido:
      type: string
      enum: [PENDIENTE, PREPARANDO, LISTO, ENTREGADO, CANCELADO]

    TipoProducto:
      type: string
      enum: [COMIDA, BEBIDA]

    LineaPedidoRequest:
      type: object
      required: [nombreProducto, cantidad, tipoProducto]
      properties:
        nombreProducto:
          type: string
          minLength: 1
        cantidad:
          type: integer
          format: int32
        tipoProducto:
          $ref: "#/components/schemas/TipoProducto"
        notas:
          type: string

    CrearPedidoRequest:
      type: object
      required: [mesa, lineas]
      properties:
        mesa:
          type: integer
          format: int32
        lineas:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/LineaPedidoRequest"

    CambiarEstadoRequest:
      type: object
      required: [estado]
      properties:
        estado:
          $ref: "#/components/schemas/EstadoPedido"

    LineaPedidoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombreProducto:
          type: string
        cantidad:
          type: integer
          format: int32
        tipoProducto:
          $ref: "#/components/schemas/TipoProducto"
        notas:
          type: string

    PedidoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mesa:
          type: integer
          format: int32
        estado:
          $ref: "#/components/schemas/EstadoPedido"
        fechaCreacion:
          type: string
          format: date-time
        fechaActualizacion:
          type: string
          format: date-time
        lineas:
          type: array
          items:
            $ref: "#/components/schemas/LineaPedidoResponse"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        mensaje:
          type: string